+++ packages/orq-rc/RELEASES.md
@@ -449,3 +449,13 @@ Based on:
 - [python v3.2.0-rc.5] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.5] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.5 - packages/orq-rc
+
+## 2025-02-17 10:49:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.493.0 (2.512.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.6] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.6] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.6 - packages/orq-rc